Initial rough proof of concept
Tracking and overview have been reimplemented using just regular Gtk,
Gdk and hamstserlib. We managed to rebuild the Overview widget from
scratch but for its fancy 'slide in' effect for the 'Totals-view'.

This is a price we are willing to pay if we get away just with standart
tools and libs. If we feel like spending time on this the right thing to
would be to implement a new 'fade-add'-container widget that povides
this functionality.

Please note, our tracking screen is rather suboptimal at this point. It
would probably be nicer to use a ``Gtk.Stack`` withour a
``StackSwitcher`` but to trigger the stack-swap when an activity is
saved/started.

From this commit onward we enter a proper project phase, no more hacking
around :)