First aproach towards testing the GUI.
We did run into severe problems using tox to handle our test
environment. For some yet unknown reason we always trigger a segfault if
our test tries to instantiate any GTK widget. At this point it is not
clear if this is an issue with ``tox`` or ``GTK``.

On top of this, there are problems running ``py34``. It apears our
system wirde GTK does miss the ``_gi`` module (C extension?).
As we do not have an alternative system to double check right now, we
don't know if this is a debian issue or something more serious.