Improve tests
Disable ``autodoc`` for docs

Until we figure out how to setup Gtk on the CI server we disable
autodocs.

There is an unidentified issue with with the linkchecker on the CI
setup. Local tests work, on the remote machine however we end up with
SSL Handshake issues. Test have been deactivated for now.