Improve Overview dialog
``OverviewScreen`` now provides a ``refresh`` method that triggers
recomputation of facts as well as a redrawing of all widgets.
For now, we go without signals and just call this method as needed.
If it turns out we actually need signals it can easily be extended.

We also changed ``hamster_gtk.HeaderBar._on_overview_button`` to reuse
an already existing dialog. However, as we ran into problems with
hiding/re-showing a once instantiated OverviewScreen class for unknown
reasons we destroy our ``OverviewScreen`` when closed entirely. As a
consequence a new one will be generated if ``_on_overview_button`` is
called again.