Add an edit dialog for existing facts
Dialog opens when a row in the overview is ``activated``.

This commit also makes sure that the ``HamsterGTK`` application is passed
along to all major players in a consistent and transparent way instead
of hoping to be able to access it as some parent window attribute.

A thing to notice is: The string used to populate the ``raw fact`` aka
``activity`` entry widget contains a proper raw fact now. That is a
legacy ``serialized_name`` prepended with the facts time info.
Another difference is that ``time frame`` is given in the proper format
where ``start`` and ``end`` are seperated by an ``-`` not by an `` `` as
in the legacy version. Whilst this breaks the previous string, it should
be alright and gets us one step closer to higher code consistency.

A new ``SignalHandler`` class is added that can be used to register
generic custom signals. This may not be the idiomatic way to do this but
then agian, is rather confusing finding any consistent information on
this matter to begin with. Instead of creating a new 'controler` class
inheriting from ``HamsterControler`` *and* ``GObject`` we opted for
``composition`` instead.  An instance of this class is added as
``HamsterGTa.controler.signal_handler`` and provides an easy way to
emit custom signals.

With regards to ``EditFactDialog.apply``:
Instead of parsing the ``raw_fact`` and tranfering the attributes to
``self._fact`` we cheat a bit and stick with the new instance and just
attach the PK of the fact to be updated before we pass it off to
``hamsterlib``.

Closes: #8