Add Dialog to pick daterange for overview(#39)
We could not implement the original design as we try not to cheat the
way it does ;) The original glade file uses a fixed width for 'today',
'week', and 'range' labels to make it apear as if there is an underlying
grid layout. However, if one wants to go with the 'wide button' design, a
proper grid layout is not an option.

Our best first aproximation now is to use ``Align.START`` and ``Align.END``
 to try to create a good locking compromise. This is most likly a place
to return to once we want to polish things. We try hard to make it as
modular and transparent as possible. So such changes should be
relatively easy if someone with some GTK mojo want to have a go.

Note:
An alternative design option would be to provide three preset button in
a titlebar.

Closes: #39