Improve passing of parents. Add basic testsuite.
This commit reduces the passing along of parent references. Instead we
focus on just passing 'app' or 'controler' references were needed.
With the exception of dialogs, most widgets can fetch their parents with
``self.get_parent``.
Because testing instantiation of all the widgets by hand became finaly
unbearable we also add a rough shot at a basic testsuite. In cases were
we just check instantiation or  a 'get_widget_*' method our test usualy
do not check for the widgets property right now but only wether or not
it successfully gets instantiated and matches the expected class.
While this maybe something that needs to be extended, this commit
provides a much needed infrastructure so that incremental improvements
can be done.

Closes: #43