Add preferencecs dialog and config handling(#16)
This commit introduces a basic preference dialog as well as basic
infrastructure to handle configuration.
A new ``config-changed`` signal has been added to the main signal
handler that is emitted whenever changes to the config are made in order
to trigger their reload.
Internally, configuration is stored as the ``HamsterGTK._config``
attribute to the main application as dictionary of config keys and
values.
Whilst in future we want to support additional config stores (GConf
comes to mind) we limit ourselfs to using ConfigParser to use config
files stored to disk. In order to do so we utilize
``hamster_lib.helpers.config_helpers`` to deal with actually reading and
writing the file. That leaves implementing methods to convert config
data between the internal config dictionary and the configparser
instance. While doing so some extra work is performed because dictionary
values are properly typed (datetime, int, etc...) while the configparser
instance only deals with strings.

In case no config file can be found, we also provide a default config.

When dealing with GTK widgets we decoded them with utf-8 as they are
returned as 'byte-strings'. This obiously led to issues under python 3
where they are returned as 'unicode text strings' and no ``decode``
method is applicable. We added a custom helper method ``_u`` that will
decode a passed ``str`` instance with utf-8 under py2 and just return
the
string under py3. Unfortunatly we were unable to use six.u for this as
it is not safe for non ascii chars (it does not make our assumptins
about the input strings encoding).

Closes: #16 #20