Add 'xvfb' to our test setup.
In order to provide a X server environment even on headless systems as
well as preventing to mess the the calling developers running X session
we uses ``xvfb`` to wrap our tests.
While this adds an editional dependency it is one that only matters for
developers, so it is a viable tradeoff.
For this to work properly within tox we also added a new maketarget
``coverage-no-xvfb`` that is identical to ``coverave`` without using
``xvfb`` (so, this is like before this commit). This make target is now
used by tox which is itself responsible for calling ``xvfb``. If ``tox``
would use the normal target ``xvfb-run`` would be called twice, causing
the second one to fail startup and hence break the tox run.

Closes: #95