Clean up Travis-Tox interation
Whilst we still not manage to run our proper py.test tests on Travis-CI
this commit makes sure that the rest of our setup is clean and well
structured. For this purpose our default tox call will now run the
entire test suite (this is what you do locally) whilst Travis-CI uses
``TOXENV`` variable to determine which environments are run. An for the
time being this includes only our 'linter' test environments.