Unset overview chart variable when switching date range
When user changed the date range in the Overview dialogue while having
chart (details) open, the chart would disappear but corresponding variable
would not be unset. Two clicks will then be required to show the chart again,
first to actually unset the variable and second to finally display the chart
widget.

This commit ensures the widget is set to False whenever the view is refreshed.

Closes: #106