Remove autodoc generated files
This commit removes accidentally added files generated by autodoc.
It also adds them to ``.gitignore`` and excludes them in ``MANIFEST.in``.

This also means that developers should not create documentation files
starting with ``hamster_gtk.`` prefix. Trouble when trying ``git add``
should alert them that it is a bad idea.

Closes: #133