Add custom 'RawFactEntry'
This commit provides a new custom entry class ``RawFactEntry`` suitable
for add/edit fact dialogs/boxes. One main feature of this new class is
that it provides basic autocomplete functionality via a custom
``Gtk.EntryCompletion`` subclass.

A major issue we had to solve was the fact that we can not simply look
up viable matches and replace the 'raw_fact' string with an activity
string because this would remove any existing time or description
informmation contained in the entry string.

Another fact worth noting is that this commit introdices serialisation
hellper methods for ``Activity`` and ``Fact`` in order to avoid code
duplication. Once ``LIB-216`` has been implemented these may no longer
be needed.

Closes: #136