Add RawFactEntry and Completion
This commit introduces a new helper ``helpers.decompose_raw_fact_string`` which is
responsible for disecting a given 'raw fact' string and extracting as
much information as possible. Unlike the ``hamster-lib`` function of
similar spirit this one takes a straight forward regex based approach to the
problem. As such it allows to return either the raw ``match`` instance
or the corresponding groupdict.

We introduce a new custom entry widget: ``RawFactEntry``. Its main
purpose is to encapsulate all raw fact related user  interaction such as
autocompletion etc. Finally we add ``RawFactCompletion`` as a custom
``Gtk.EntryCompletion`` class that is responsible for actually providing
autocompletion.

Closes: #148