'activits+catagories' can match against just categories
If the decomposed raw fact does not include an 'activity' but a
'category' segment (e.g. "@foo") 'activity+category' will now be able to
deal with this and as a consequence will autocomplete against all
activities with this category. This allows for fast activity lookup if
the category is known but not the activity name.