Pass app config to RawFactCompletion
In order to make the apps config available for the ``RawFactCompletion``
in our ``StartTracking`` screen we have to pass it thorugh all relevant
parrents. We do so explicitly and do not simply pass the app as a whole
in oder to be more transparent about what we do and do not need.
This also means adjusting some tests/fixtures.