overview: Refactor overview dialog
This commits improves and extends our OverviewDialog with a focus  on
the 'details' area.

code improvements:
- Move ``get_delta_string`` to a dedicated helper funciton.
- When computing ``totals`` for the overview dialogs we previously
provided all deltas for the relevant key in the ``*_deltas`` dicts.
This seemes unwarented as we just care for the total value per key. So the
new solution does just that. Instead of a list the value is now one
``datetime.timedelta`` instance that represent the accumulated deltas
for this key. This caused a few changes in the consuming code.

GUI improvements:
- We extend the overview details panel to list activities and dates as
well. While doing so we improved styling by reducing fontcolour,
providing coloumn headers and adjusting barsize.
- Barcharts are not order by the ``delta`` they represent in their given
'category'.