Exclude tests from installation
Previously, the tests module was included when installing the application.
This could lead to a conflict when another python package also exposed its
tests and used a same file, for example `tests/helpers/__init__.py`.

This patch excludes the tests module from installation. The tests module
will still be included in the distribution tarballs so that hamster-gtk
can be tested before installation.

Closes: #189