Fix 'app' fixture
Our ``app`` fixture (which is used extensively) used to instantiate the
default ``HamsterGTK`` class. This meant that the default config loading
was triggered and as a consequence the systems live config being used.
This is obviously an issue. Instead we monkey patch the class now to
skip the config loading machinery and use/return the ``config`` fixture.